+++ noodles-cram/src/reader.rs
@@ -344,7 +344,7 @@ where
             self.records(reference_sequence_repository, header)
                 .map(|result| {
                     result
-                        .and_then(|record| record.try_into_sam_record(header))
+                        .and_then(|record| record.try_into_alignment_record(header))
                         .map(|record| Box::new(record) as Box<dyn sam::AlignmentRecord>)
                 }),
         )